requirements.txt
torch
torchvision
numpy
pandas
matplotlib
scikit-learn
Pillow
tqdm